Hey guys! Welcome to a series that we're calling: Journey to the Center of Symfony! In this first part, we'll be talking about the deep, dark core piece called the HttpKernel, a wondrous component that not only sits at the heart of Symfony, but also at the heart of Silex, Drupal 8, PhpBB and a lot of other stuff. How is that possible? We'll find out! And this stuff is really nerdy, so we're going to have some fun.
Big Picture: Request-Route-Controller-Response

No matter what technology or framework we're using, our goal is always to start with a request and use that to create a response. Everything in between those 2 steps will be different based on your tech or framework. In our app, and in almost every framework, two things that are going to be between the request and response are the route and controller. In this case, you can see our homepage has a route, our function is a controller, and our controller returns a Response object:

42 lines  src/AppBundle/Controller/DinosaurController.php
... lines 1 - 10
  /**
   * @Route('/', name='dinosaur_list')
   */
  public function indexAction(){
    {dollar_sign}dinos = {dollar_sign}this->getDoctrine()
     ->getRepository('AppBundle:Dinosaur')
     ->findAll();
    return {dollar_sign}this->render('dinosaurs/index.html.twig', [
      'dinos' => {dollar_sign}dinos,
    ]);
  }
... lines 24 - 42

And we have the same thing down here with the other page: it has a route, a controller, and that returns a response::

 42 lines  src/AppBundle/Controller/DinosaurController.php
... lines 1 - 24
  /**
   * @Route('/dinosaurs/{id}', name='dinosaur_show')
   */
  public function showAction({dollar_sign}id){
    {dollar_sign}dino = {dollar_sign}this->getDoctrine()
      ->getRepository('AppBundle:Dinosaur')
      ->find({dollar_sign}id);
      throw ->createNotFoundException('That dino is extinct! ');
    }
    return {dollar_sign}this->render('dinosaurs/show.html.twig', [
      'dino' => {dollar_sign}dino,
    ]);
  }
So what we're going to look at is how that all works. Who actually runs the router? Who calls the controller? How do events work in between the request-response flow?

But before we dive into that, what we're going to do first is create an event listener and hook into that process. Then we'll be able to play with that event listener as we dive into the core of things.