Logging Something in the Subscriber
Now inside of onKernelRequest, let's do some real work. For now, I want to log a message. I'm going to need the logger so I'll add a constructor and even type hint the argument with the PSR LoggerInterface. And I'll use a little PHPStorm shortcut to create and set that property for me.

Now in our function, we'll log a very important message.

And of course this isn't going to work unless we go back to services.yml and tell Symfony: Hey, we need the @logger service.

Let's refresh! It works, and if we click into the profiler, one of the tabs is called Logs, and under info we can see the message. So this is already working, and if we go back to the Timeline and look closely, we should see our UserAgentSubscriber. And it's right there. Also, if we go back to the events tab, we see the kernel.request with all of its listeners. And if you look at the bottom, you see our UserAgentSubscriber on that list too.

So we're hooking into that process already, even if we don't understand what's going on with it.