HttpKernel::handle() The Heart of Everything We know we start with the request, we have a routing layer, eventually something calls the controller, and the controller returns a Response. Let’s trace through the process in Symfony to see how all of this works. It’s going to be awesome, I promise.
Front Controller: app.php/app_dev.php

Since we’re using the built-in web server, this is actually executing the app_dev.php file in the web/ directory. So let’s go there to start. And really, app.php is basically identical, so it doesn’t matter which one we start going through. Ignore all the stuff at the top of app_dev.php, I want to get to the good stuff

The first important thing is that we instantiate the AppKernel, which is the same AppKernel you have in your app/ directory where you register your bundles. We’ll talk more about this class in another part of this series. We’ll also talk then about this loadClassCache, but for now, comment it out - it can get in the way if you’re debugging core classes.

For now, just know that AppKernel is the heart of your app. The first real important thing I want to look at is the Request::createFromGlobals. We all know that we start with the request, and hey, here’s where that Request is born! The Request is just a simple object that holds data, and this creates a new one that’s populated from the superglobal variables(http://php.net/manual/en/language.variables.superglobals.php) - those things like {dollar_sign}_GET, {dollar_sign}_POST, {dollar_sign}_SERVER. In fact, if we look inside, it’s creating an instance of itself populated from those variables:

// vendor/symfony/symfony/src/Symfony/Component/HttpFoundation/Request.php
// ...

public static function createFromGlobals(){
  {dollar_sign}request = self::createRequestFromFactory({dollar_sign}_GET, {dollar_sign}_POST, array(), {dollar_sign}_COOKIE, {dollar_sign}_FILES, {dollar_sign}_SERVER);
  // ...

  return {dollar_sign}request;
}

The Most Important Function in Symfony

Now that we’re starting with the request object, the next line is the most important line inside of Symfony: the ->handle() method:

// web/app_dev.php
// ...

{dollar_sign}response = {dollar_sign}kernel->handle({dollar_sign}request);
// ...
What’s really cool about this is that you see that the handle() method has input request, output response. And that is actually amazing. It means that everything that our application is - the routing, the controller, the services - all of that stuff happens inside of this handle() function. It also means that our entire application - all of those layers - have been boiled down to a single function. And how pretty is it: input request, output response. Because after all, that’s our job: read the incoming request, create a response.