Why Yes, you can Handle Many Requests at Once
What I want to do with you guys is look inside that handle() function and see what’s going on. But first, I kind of want to prove a point about how our application has just been reduced down to a single function. One of the things about a well-written function is that you can call it over and over again. So in theory, if we wanted to here, we could actually create a second request by hand. I’ll use a create() function

What I want to do is process two requests at once. As you’ll see later, internally in Symfony, there actually is a real use-case for this. But for now, I’m just playing around, so we’ll have this request look like it wants to go to /dinosaurs/22, which is.... one of the pages we actually have right now.

Then we’ll say {dollar_sign}response2 = {dollar_sign}kernel->handle({dollar_sign}request2);. We’re processing two separate requests and processing two separate responses. I’m going to comment out the send() and terminate() calls temporarily, and to prove that things are working, I’m going to echo the content from the original request, then echo the content from the second response. So hey, let’s see if we can print out two pages at once.

When we go back and refresh, this is really cool! On top, we see page that we’re actually going to, and below, we see the whole other page that was processed beneath that. Our application is just a function: input request, output response. And that’s a really powerful thing to realize.