Introducing HttpKernel::handle()
Let’s look inside of that ->handle() method. Again, the  class is our AppKernel. If I hold cmd (or ctrl for other OS’s) and click into the handle() function, it’s going to take us not into AppKernel, but its parent class Kernel. This class is something we’re going to talk about in a different part of this series. Ignore it for now, because it offloads the work to something called HttpKernel.

I’ll use a shortcut my editor to open HttpKernel. In PhpStorm, you can go to Navigate, then Class or File. So I’ll use the Cmd+O shortcut to open up the HttpKernel class. We’re looking for that handle() function. Because effectively, when we call handle() in app_dev.php, it’s being passed to this handle() function - you can see the  argument:

// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/HttpKernel.php
// ...

public function handle(Request {dollar_sign}request, {dollar_sign}type = HttpKernelInterface::MASTER_REQUEST, {dollar_sign}catch = true){
  try {
    return {dollar_sign}this->handleRaw({dollar_sign}request, {dollar_sign}type);
  } catch (\Exception {dollar_sign}e) {
    if (false === {dollar_sign}catch) {
      {dollar_sign}this->finishRequest({dollar_sign}request, {dollar_sign}type);

        throw {dollar_sign}e;
      }

      return {dollar_sign}this->handleException({dollar_sign}e, {dollar_sign}request, {dollar_sign}type);
  }
}
Awesome! Now, what is handle actually doing? So far, not much. The important thing to take-away here is that there is a try-catch block. This means that if you throw an exception from anywhere inside your application - like a controller or a service - it’s going to get caught by this block. And when that happens, you’ll get passed to that handleException() function, which is what tries to figure out what response to send back to the user when there’s an error. That’s something we’ll talk about later.

The real guts of this are in a function called handleRaw(). And this lives just a little bit further down inside this same class:

private function handleRaw(Request {dollar_sign}request, {dollar_sign}type = self::MASTER_REQUEST){
    // about 45 lines of awesome that we'll walk-through
}
We’re going to walk through every line in this function. You can see that it’s not that long, and what’s really amazing is that handleRaw is the Symfony Framework. This is the dark core guts of it. But it’s also the dark, core guts of Drupal 8, and the dark, core guts of Silex, of PhpBB. All these very different pieces of software all use this same function. How is that possible? How could this one function be responsible for a Symfony application and a Drupal 8 application and a PhpBB application? We’ll find out.