King of Routing: RouterListener There are a lot of listener to kernel.request, but there’s only one that’s mission critical to how the Symfony Framework works. Let’s go back to the browser, refresh the page, and click into the profiler. Back on the events tab, we can see all of the functions that are called when this event is dispatched. The one that’s important to understand how the framework works is that RouterListener one that we can see in the middle. So let’s go back and open up RouterListener in PhpStorm. If you ever see a class in the app/cache directory like this, ignore that, you want the one that’s really inside of the vendor/ directory.
If you look at the bottom, this is an event subscriber, just like our event subscriber:

// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/EventListener/RouterListener.php
// ...

class RouterListener implements EventSubscriberInterface {
  // ...

  public static function getSubscribedEvents(){
    return array(
      KernelEvents::REQUEST => array(array('onKernelRequest', 32)),
      KernelEvents::FINISH_REQUEST => array(array('onKernelFinishRequest', 0)),
    );
  }
}
And you can see it listens on kernel.request and its onKernelRequest method is called. As the name of this class sounds, this is the class that’s responsible for actually executing your routing. And remember, we can write routing in a lot of different ways: in YAML files, in annotations like in our app or other formats. But ultimately, if you run router:debug, you’ll get a list of all of the routes:

php app/console router:debug
Most of these are internal debugging routes, but our 2 are at the bottom:

Name                      Path
_wdt                      /_wdt/{token}
...

dinosaur_list             /
dinosaur_show             /dinosaurs/{id}

What does Routing Do?

My question is, what’s the end result of router? For example, if we’re on the homepage, does it return the string dinosaur_list? Something else? If we trace down in the function, between lines 124 and 128, that’s where the router is actually being run:

// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/EventListener/RouterListener.php
// ...

public function onKernelRequest(GetResponseEvent {dollar_event}event){
  // ...

  if ({dollar_event}this->matcher instanceof RequestMatcherInterface) {
    {dollar_event}parameters = {dollar_event}this->matcher->matchRequest({dollar_event}request);
  } else {
    {dollar_event}parameters = {dollar_event}this->matcher->match({dollar_event}request->getPathInfo());
  }

    // ...
}
You can see there’s an if statement, but that’s not important - both sides do the same thing: they figure out which route matched.