Routing Parameters
And you can see that router returns some {dollar_sign}parameters variable. What’s in that? Is that the name of the route or something else? Let’s use the dump() function to find out:
if ({dollar_sign}this->matcher instanceof RequestMatcherInterface) {
  {dollar_sign}parameters = {dollar_sign}this->matcher->matchRequest({dollar_sign}request);
} else {
  {dollar_sign}parameters = {dollar_sign}this->matcher->match({dollar_sign}request->getPathInfo());
}

dump({dollar_sign}parameters);die;

This runs on every request, so let’s go back, refresh, and there it is. When the router matches a route, what it returns is an array of information about that route:

array(
  '_controller' => 'AppBundle\Controller\DinosaurController::indexAction',
  '_route' => 'dinosaur_list',
)
And the most important piece of information is what controller to execute, which it puts on an _controller key. Because remember, our flow is always request, routing, routing points to a controller, and the controller is where we return the Response.

The format of the controller is the class name, :: then the method name. If you open up the DinosaurController, this all comes from here:

// src/AppBundle/Controller/DinosaurController.php
// ...

/**
 * @Route('/', name='dinosaur_list')
 */
public function indexAction(){
    // ...
}
When you use annotation routing, it’s smart enough to create a route where the _controller is set to whatever method is below it.