The _controller key and Yaml Routes
If you’ve used YAML routes before, it’s even more interesting. I’ll open up my app/config/routing.yml file. At the bottom in comments, I prepared a route that’s identical to the one we’re building in annotations.

When you use YAML routing, to point to the controller you have a defaults key, and below that you have an _controller that uses a three-part syntax to point to the DinosaurController and indexAction.

So just to prove this is the same. I’m going to load only that route.

And when we refresh, the routing parameters are exactly the same.