Routing Wildcards are Parameters Too!
Now let’s go to a different page - /dinosaurs/22. And what we see now is that in addition to _controller and _route, now we have the id value from the route:

array(
  '_controller' => 'AppBundle\Controller\DinosaurController::showAction',
  'id' => '22',
  '_route' => 'dinosaur_show',
)
This shows that the end result of the routing layer is an array that has the _controller key plus any wildcards that are in your URL. It also gives you the _route in case you need it, but that’s not important.