Let’s try this. In UserAgentSubscriber, I’ll create a new Response object. Make sure you use the one from the HttpFoundation component. PHPStorm did just add a use statement. And we’ll say 'Come back later'
And the GetResponseEvent object we’re passed has an {dollar_sign}event->setResponse() method on it. Remember, every event passes a different object, and this one happens to have this setResponse method on it. We’ll make things more interesting and put this in a block so it only randomly sets the response.

If we go and refresh now, the page works fine, refresh again... mine is being stubborn. There we go. You can see that as soon the response is set, it just stops entirely.