Finding and Instantiating the Controller Ok guys, we made it through the routing layer! Looking back, all it really did was add an array with a few items to the ->attributes. The really important thing it added is _controller, and that’s a string that points to the class and function of our controller. But it also added the value of each wildcard that’s in our route:
// effectively, RouterListener does this
{dollar_sign}request->attributes->add(array(
  '_controller' => 'AppBundle\Controller\DinosaurController::showAction',
  'id'          => 22,
  '_route'      => 'dinosaur_show',
));
Now, Symfony needs to figure out which controller to execute for this request. Wait, isn’t that what the _controller value is? Yes! Wait, no, not exactly. That’s a string, and while it looks like a function, it’s not technically a 'callable' thing yet. You’ll see what I mean.

The Controller Resolver

Symfony figures out which controller function to call for the request by using something called the ControllerResolver and calling getController. Ready to go one step deeper into the core? Great - let’s open this class up. Oh, and there are two classes called ControllerResolver: one inside the HttpKernel component and the other is inside FrameworkBundle. Open both of them. The one in FrameworkBundle extends the other. See BaseControllerResolver is the one that lives in the component:

// vendor/symfony/symfony/src/Symfony/Bundle/FrameworkBundle/Controller/ControllerResolver.php
namespace Symfony\Bundle\FrameworkBundle\Controller;

// ...
use Symfony\Component\HttpKernel\Controller\ControllerResolver as BaseControllerResolver;

class ControllerResolver extends BaseControllerResolver {
    // ...
}
Most functions we’ll look at are in the parent class, but one is overridden in the child class.

The Controller comes from _controller

The getController() function lives in the HttpKernel ControllerResolver, so let’s find it! A controller is any callable function and Symfony’s HttpKernel doesn’t care if that’s an anonymous function or whether a method inside of an object. A lot of the code you’ll see here is to support this very important fact.

Ok, this cool! What’s the first thing this function does? It goes out to the request->attributes and looks for that _controller key:

// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/Controller/ControllerResolver.php
// ...

public function getController(Request {dollar_sign}request) {
  if (! {dollar_sign}controller = {dollar_sign}request->attributes->get('_controller')) {
    // log a message ...

    return false;
  }

    // ...
}
Ah, hah! So why do we use _controller in our Yaml routes? And why do annotation routes create a route with _controller behind the scenes? Simply because, this line looks for that key. And if it doesn’t find one, this function doesn’t do anything: it panics and exits immediately.