Is the _controller Already a Callable?
Most of the rest of the code is basically trying to figure out: 'Hey, is the _controller key in the request attributes maybe already a callable function?' In our case it’s not, and let’s dump() it to get a reminder of what it looks like at this point:

/ vendor/symfony/symfony/src/Symfony/Component/HttpKernel/Controller/ControllerResolver.php
// ...

public function getController(Request {dollar_sign}request){
  if (! {dollar_sign}controller = {dollar_sign}request->attributes->get('_controller')) {
    // log a message ...

    return false;
  }

  dump({dollar_sign}controller);die;
    // ...
}

Go back to the homepage, refresh!

(string) 'AppBundle\Controller\DinosaurController::indexAction'
Our controller is this string, not technically a callable function yet, even though it looks like a function name. A variable is callable if you can invoke it via the call_user_func function.

But in other circumstances, the controller might already be callable. Silex is a perfect example! There, your controllers are anonymous functions, and behind the scenes, the function is set on the _controller key of the route:

// example silex app
{dollar_sign}app = new Silex\Application();
{dollar_sign}app->get('/dinosaurs/{id}', function({dollar_sign}id) {
    return 'ROAR Dinosaur #'.{dollar_sign}id;
});

// this is what happens in RouterListener
{dollar_sign}request->attributes->add(array(
  '_controller' => function({dollar_sign}id) {
     return 'ROAR Dinosaur #'.{dollar_sign}id;
  },
  'id' => 22,
));
So for Silex, _controller is callable, and so it would exit earlier in this process.