Transforming _controller into a Callable
But in the Symfony Framework, we don’t exit early. Instead, we fall down into the createController function:

// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/Controller/ControllerResolver.php
// ...

public function getController(Request {dollar_sign}request){
  // ...

  {dollar_sign}callable = {dollar_sign}this->createController({dollar_sign}controller);
  // ..

  return {dollar_sign}callable;
}
This is overridden in the child ControllerResolver, so switch to the one that’s in the FrameworkBundle. And there’s our function!

// vendor/symfony/symfony/src/Symfony/Bundle/FrameworkBundle/Controller/ControllerResolver.php
// ...

protected function createController({dollar_sign}controller){
  if (false === strpos({dollar_sign}controller, '::')) {
    // ...
  }

    // ...
}

Ah, this is awesome again! Look at the first part: it looks to see if your controller already has a :: syntax in the middle of it. And if you don’t you fall into this first block.

Transforming AppBundle:Default:index

Remember that AppBundle:Default:index syntax you use in Yaml routing? That’s handled here. The ->parser you see on line 60 is responsible for transforming that 3-part syntax and into the longer class name :: method name that our’s already has:

// vendor/symfony/symfony/src/Symfony/Bundle/FrameworkBundle/Controller/ControllerResolver.php
// ...

protected function createController(){
  if (false === strpos({dollar_sign}controller, '::')) {
    // ...

    // controller in the a:b:c notation then
    {dollar_sign}controller = {dollar_sign}this->parser->parse({dollar_sign}controller);

    // ...
  }
  // ...
}
Tip
In reality, there is a similar layer that runs during the process of compiling routes that converts the AppBundle:Default:index syntax to AppBundle\Controller\DefaultController::indexAction. But the idea is still the same - this just makes runtime performance a bit better.

The Controller as a Service Syntax

The block below this - around line 63 - handles the service syntax (e.g. my_dinosaur_controller:indexAction:

// vendor/symfony/symfony/src/Symfony/Bundle/FrameworkBundle/Controller/ControllerResolver.php
// ...

protected function createController({dollar_sign}controller){
  if (false === strpos({dollar_sign}controller, '::')) {
    // ...

    // controller in the service:method notation
    list({dollar_sign}service, {dollar_sign}method) = explode(':', {dollar_sign}controller, 2);

    return array(->container->get(), );

    // ...
  }
  // ...
}

So if you decide to register your controller as a service, you have a different syntax, and this is the logic that handles that.

Ultimately, we fall down to the bottom, and one way or another, we end up with a string, which is the class name, ::, and then the method name.

Next, this splits those on the :: and the strange list() function sets the first part to a variable called  and everything after the :: to a variable called . I’ll dump those variables to be totally clear:

// vendor/symfony/symfony/src/Symfony/Bundle/FrameworkBundle/Controller/ControllerResolver.php
// ...

protected function createController({dollar_sign}controller) {
  // ...

  list({dollar_sign}class, {dollar_sign}method) = explode('::', {dollar_sign}controller, 2);
  dump({dollar_sign}class, {dollar_sign}method);die;
  // ...
}

The list() function has confused me in the past, and the real key is what these two new variables are set to.

At this point, it’s time to see if we’ve messed up! Maybe this class doesn’t exist - maybe there’s a typo somewhere. It gives us a nice error message in that case.