Injecting the Container (ContainerAwareInterface)
The next line is really really important to just about everything you do every day in Symfony. It says: if your controller object implements the ContainerAwareInterface, then call {dollar_sign}controller->setContainer({dollar_sign}container):

// vendor/symfony/symfony/src/Symfony/Bundle/FrameworkBundle/Controller/ControllerResolver.php
// ...

protected function createController({dollar_sign}controller) {
  // ...

  {dollar_sign}controller = new {dollar_sign}class();
  if ({dollar_sign}controller instanceof ContainerAwareInterface) {
    {dollar_sign}controller->setContainer({dollar_sign}this->container);
  }

  return array({dollar_sign}controller, {dollar_sign}method);
}
So if I open up DinosaurController and click to open Symfony’s base Controller, you’ll see that it extends a ContainerAware class:

namespace Symfony\Bundle\FrameworkBundle\Controller;

use Symfony\Component\DependencyInjection\ContainerAware;
// ...

class Controller extends ContainerAware
{
    // ...
}
Let’s click to open that. And we see that it implements the ContainerAwareInterface:

// vendor/symfony/symfony/src/Symfony/Component/DependencyInjection/ContainerAware.php
namespace Symfony\Component\DependencyInjection;

abstract class ContainerAware implements ContainerAwareInterface {
  protected {dollar_sign}container;

  public function setContainer(ContainerInterface {dollar_sign}container = null){
    {dollar_sign}this->container = {dollar_sign}container;
  }
}

So if our controller extends Symfony’s base Controller, we automatically implement that interface. Because of that, the ControllerResolver does call setContainer on our controller class, which is this function here. And what does it do? It sets that on a protected  property. And this is the reason why in any controller function, we can say {dollar_sign}this->container->get() and then get out whatever service we want:

public function indexAction(){
    {dollar_sign}this->container->get('logger')->alert('DINOS, RUN!');
}
If, for some reasons, you didn’t want to extend Symfony’s base Controller, but still wanted access to the container, that would be fine: you’d just need to implement that ContainerAwareInterface and then have, maybe, a similar setContainer method that sets it on a  property.

Back in ControllerResolver, we now have a {dollar_sign}controller object, we have the {dollar_sign}method that’s going to be called on it, and it returns an array with those two things:

// vendor/symfony/symfony/src/Symfony/Bundle/FrameworkBundle/Controller/ControllerResolver.php
// ...

protected function createController({dollar_sign}controller){
  // ...

  {dollar_sign}controller = new {dollar_sign}class();
  if ({dollar_sign}controller instanceof ContainerAwareInterface) {
    {dollar_sign}controller->setContainer({dollar_sign}this->container);
  }

  return array({dollar_sign}controller, {dollar_sign}method);
}
This is a 'callable' format syntax in PHP. This ultimately goes back to the other ControllerResolver and is returned all the way back to HttpKernel::handleRaw().