kernel.controller Event & Controller Arguments Ok guys, what next? Hey, let’s dispatch another event:
// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/HttpKernel.php
// ...

private function handleRaw(Request {dollar_sign}request, {dollar_sign}type = self::MASTER_REQUEST){
  // ...

  {dollar_sign}event = new FilterControllerEvent({dollar_sign}this, {dollar_sign}controller, {dollar_sign}request, {dollar_sign}type);
  {dollar_sign}this->dispatcher->dispatch(KernelEvents::CONTROLLER, {dollar_sign}event);
  {dollar_sign}controller = {dollar_sign}event->getController();

  // ...
}
This one is called kernel.controller. And like I keep promising, this one is passed a totally different event object called FilterControllerEvent. Why would you listen to this event? Well, FilterControllerEvent has a getController() method on it. So if you needed to do something based on what the controller is, or even change the controller in a listener, maybe to mess with your co-workers, a listener to this can do that.

But for understand the Symfony Framework, this is just a hook point. There are no mission-critical listeners to this event. So let’s keep moving!