Finding Values for Each Argument
In doGetArguments(), the first thing it does is so important and so geeky cool. It goes back to ->attributes - that same thing that was populated by the routing. Let’s dump this out quickly to remember what’s in there:

// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/Controller/ControllerResolver.php
// ...

protected function doGetArguments(Request {dollar_sign}request, {dollar_sign}controller, array {dollar_sign}parameters){
  {dollar_sign}attributes = {dollar_sign}request->attributes->all();
  dump({dollar_sign}attributes);die;

  // ...
}
When we refresh, it has _controller and id from the routing wildcard:

array(
  '_controller' => 'AppBundle\Controller\DinosaurController::showAction',
  'id'          => '22',
  '_route'      => 'dinosaur_show',
  '_route_params' => array(...),
)
It also has a couple of other things that honestly aren’t very important.

Keep that array in mind. The doGetArguments() function iterates over : the array of info about the arguments to our controller. And the first thing it does is check to see if the name of the parameter - like id - exists in the {dollar_sign}attributes array. And if it does, it uses that value for the argument:

// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/Controller/ControllerResolver.php
// ...

protected function doGetArguments(Request {dollar_sign}request, {dollar_sign}controller, array {dollar_sign}parameters){
  {dollar_sign}attributes = {dollar_sign}request->attributes->all();
  {dollar_sign}arguments = array();

  foreach ({dollar_sign}parameters as {dollar_sign}param) {
    if (array_key_exists({dollar_sign}param->name, {dollar_sign}attributes)) {
      {dollar_sign}arguments[] = {dollar_sign}attributes[{dollar_sign}param->name];
    } elseif {
      // ...
    }
  }

  return {dollar_sign}arguments;
}
This is exactly why if we have a {id} inside our route, its value is passed to a  argument in the controller. This is why and how that mapping is done by name, because ControllerResolver say: 'Hey, I have an {dollar_sign}id argument, is there an id inside the {dollar_sign}request->attributes, which is populated by the routing?'