The Request Argument And if there is nothing in the attributes that matches up, it goes to the elseif: because there’s one other case. And you’ve probably seen it while doing form processing. This is when you have an argument type-hinted with the Request class:
// src/AppBundle/Controller/DinosaurController.php
// ...
use Symfony\Component\HttpFoundation\Request;

/**
 * @Route('/dinosaurs/{id}', name='dinosaur_show')
 */
public function showAction({dollar_sign}id, Request {dollar_sign}request)
{
    // ...
}
In ControllerResolver, it says: if ({dollar_sign}param->getClass() && {dollar_sign}param->getClass()->isInstance({dollar_sign}request)):

// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/Controller/ControllerResolver.php
// ...

protected function doGetArguments(Request {dollar_sign}request, {dollar_sign}controller, array {dollar_sign}parameters){
   = ->attributes->all();
   = array();

  foreach ({dollar_sign}parameters as {dollar_sign}param) {
    if (array_key_exists({dollar_sign}param->name, {dollar_sign}attributes)) {
      {dollar_sign}arguments[] = {dollar_sign}attributes[{dollar_sign}param->name];
    } elseif ({dollar_sign}param->getClass() && {dollar_sign}param->getClass()->isInstance({dollar_sign}request)) {
      {dollar_sign}arguments[] = {dollar_sign}request;
    }
    // ...
  }

  return {dollar_sign}arguments;
}
The {dollar_sign}param->getClass() asks if the argument has a type-hint. The second part checks to see if the type-hint is for Symfony’s Request class. If it is, the {dollar_sign}request object is passed to this argument. This is completely special to the Request object: it doesn’t work with anything else.

But guys, it’s really pretty if you think about it. Our whole job is to read the request and create the response. So, It makes a lot of sense to be able to have the Request as an argument to a function that will return the Response. Input Request, output Response.