The kernel.view Event We’ve journeyed far, brave traveler. And now that Symfony has called our controller, let’s start our adventure home.
Our controller is returning a Response object. We’re calling render() and that’s a shortcut that renders the template and put it inside of a Response object:

// src/AppBundle/Controller/DinosaurController.php
// ...

public function showAction({dollar_sign}id){
  // ...

  return {dollar_sign}this->render('dinosaurs/show.html.twig', [
    'dino' => {dollar_sign}dino,
  ]);
}
And even though I love to say that the one job of your controller is to create and return a Response, it’s a lie! You can return whatever you want from a controller.

And if you don’t return a response, what does Symfony do? It dispatches another event:

// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/HttpKernel.php
// ..

private function handleRaw(Request {dollar_sign}request, {dollar_sign}type = self::MASTER_REQUEST){
  // ...

  // call controller
  {dollar_sign}response = call_user_func_array({dollar_sign}controller, {dollar_sign}arguments);

  if (! {dollar_sign}response instanceof Response) {
    {dollar_sign}event = new GetResponseForControllerResultEvent({dollar_sign}this, {dollar_sign}request, {dollar_sign}type, {dollar_sign}response);
    {dollar_sign}this->dispatcher->dispatch(KernelEvents::VIEW, {dollar_sign}event);

    if ({dollar_sign}event->hasResponse()) {
      {dollar_sign}response = {dollar_sign}event->getResponse();
    }

        // ...
  }

  // ...
}
This time it’s called kernel.view. The purpose of a listener to this event is to save Symfony. It really needs a Response object, and if the controller didn’t give it to us, it calls each listener to this event and says: 'Hey, here’s what the controller did return, can you somehow transform this into a real response and save the day?'

Ok, but what’s the use-case for this? In a true MVC framework, the controller is supposed to just return some data, not a response. For example, imagine if our DinosaurController::showAction just returned a Dinosaur object instead of the HTML response. And then, imagine we added a listener on this event that saw that Dinosaur object, rendered the template and created the Response for us.

The result would be the same, but we’d be splitting things into two pieces. The fetching and preparation of the data would happen in the controller. The creation of a representation of that data would happen in the listener.

To some of you, this might sound ridiculous. I mean, why do all this extra work? One real use-case is for APIs. Imagine the endpoints of your API need to be able to return both HTML or JSON depending on the Accept header sent by the client. In your controller, instead of having a big if statement for the two different formats, it just return the data:

// src/AppBundle/Controller/DinosaurController.php
// ...

public function showAction({dollar_sign}id){
  {dollar_sign}dino = {dollar_sign}this->getDoctrine()
    ->getRepository('AppBundle:Dinosaur')
    ->find({dollar_sign}id);

   // ...

   return {dollar_sign}dino;
}
Then you’d have a listener on kernel.view that first checks to see if the user wants HTML or JSON. If the user wants HTML, it would render the template. If it wants JSON, it would take that Dinosaur object and turn it into JSON:

// some imaginary listener class

public function onKernelView(GetResponseForControllerResultEvent {dollar_sign}event){
  {dollar_sign}format = {dollar_sign}this->checkAcceptHeader({dollar_sign}event->getRequest());
  {dollar_sign}dino = {dollar_sign}event->getControllerResult();
  if ({dollar_sign}format == 'html') {
    {dollar_sign}html = {dollar_sign}this->templating->render('dinosaurs/show.html.twig', [
      'dino' => {dollar_sign}dino,
     ]);

      return new Response({dollar_sign}html);
    } elseif ({dollar_sign}format == 'json') {
      {dollar_sign}json = {dollar_sign}this->serializeToJson({dollar_sign}dino);

      return new Response({dollar_sign}json, 200, [
        'Content-Type' => 'application/json'
      ]);
    }
}

The FOSRestBundle(https://github.com/FriendsOfSymfony/FOSRestBundle/blob/master/Resources/doc/3-listener-support.rst#view-response-listener) has something that does exactly this: you can return data from your controller, and it has a listener on the kernel.view event that transforms that data into whatever response is asked for.

Tip
If you register controllers as services, there’s an additional use-case for this. See Lightweight Symfony2 Controllers(http://www.whitewashing.de/2014/10/14/lightweight_symfony2_controllers.html).

In normal Symfony, there’s nothing important that listens to this event. But when it’s dispatched, it’s hoping that one of those listeners will be able to set the response on the {dollar_sign}event object. If we still don’t have a Response, that’s when you get the error that the controller must return a Response. Oh, and this is one of my favorite parts of the code: if the {dollar_sign}response is actually null, and it says 'Hey, maybe you forgot to add a return statement somewhere in your controller?':

// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/HttpKernel.php
// ..

private function handleRaw(Request {dollar_sign}request, {dollar_sign}type = self::MASTER_REQUEST){
  // ...

  if (! {dollar_sign}response instanceof Response) {
    // dispatch the event
    // ...

    if ({dollar_sign}event->hasResponse()) {
      {dollar_sign}response = {dollar_sign}event->getResponse();
    }

    if (! {dollar_sign}response instanceof Response) {
      {dollar_sign}msg = sprintf('The controller must return a response (%s given).', {dollar_sign}this->varToString({dollar_sign}response));

      // the user may have forgotten to return something
      if (null === {dollar_sign}response) {
        {dollar_sign}msg .= ' Did you forget to add a return statement somewhere in your controller?';
      }
      throw new \LogicException({dollar_sign}msg);
    }
  }

    // ...
}
Yep, I’ve seen that error a few times in my days.

Since our controller is returning a Response, if we go back and look at the events tab in the profiler, we’ll see that there’s no kernel.view in the list at the top. But below there’s a 'Not Called Listeners' section, and there is one listener to kernel.view, which comes from the SensioFrameworkExtraBundle, that’s not being executed.